Two Player Reaction Timer
2 player game as follows:

When a start button is pressed a 3 round game is initiated and the buzzer should sound after a random number of seconds (<10)

When the buzzer sounds, each player reacts by pressing a button. The reaction time difference should be displayed on the LCD. In addition, the round winner should be indicated by a green LED and the loser by a red LED.

The game should continue with the buzzer sounding after a random  number of seconds. After 3 rounds have been completed, the overall winner should be displayed on the LED together with the total time difference between the winner and loser.

If a player presses their button before the buzzer sounds in any round, they should be reported as the loser and the game should terminate.